\name{example_g_pars}
\docType{data}
\alias{example_g_pars}
\alias{example_h_pars}
\alias{example_h_2_pars}
\title{Genotype data parameters}
\description{
A small example of genetic data parameters.
}
\usage{example_g_pars}
\format{
A data frame of 21 parameters. The name and the value.
}
\references{Burkett KM, McNeney B, Graham J. Sampletrees and Rsampletrees: sampling gene genealogies conditional on SNP genotype data. Bioinformatics. 32:1580-2, 2016}
\keyword{datasets}
