\name{Rsampletrees-package}
\alias{Rsampletrees-package}
\alias{Rsampletrees}
\docType{package}
\title{
Pre and Post-Processing of sampletrees output
}
\description{
Contains functions that are useful for pre- and post-processing the tree output from the MCMC-based C++ program sampletrees. 
}
\details{
\tabular{ll}{
Package: \tab Rsampletrees\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014/08/01\cr
License: \tab GPL-2\cr
}

Use the pre-processing functions to create and write settings values to a file.

After running sampletrees, use post-processing functions to read the sampletrees output (acceptance rates, mutation and recombination rates, trees) into R in order to summarize the results.

Future versions will have a sampletrees function that calls the C++ code directly.

}
\author{Kelly Burkett}
\references{TBD}

