\name{Rsampletrees-package}
\alias{Rsampletrees-package}
\alias{Rsampletrees}
\docType{package}
\title{
MCMC sampling of gene genealogies conditional on genetic data
}
\description{
Sample ancestral trees conditional on phased or unphased SNP genotype data. The actual tree sampling is done using a C++ program that is launched within R. The package also contains functions for specifying the tree-sampling settings (pre-processing) and for storing and manipulating the sampled trees (post-processing). More information about sampletrees can be found at http://stat.sfu.ca/statgen/research/sampletrees.html. 

}
\details{
\tabular{ll}{
Package: \tab Rsampletrees\cr
Type: \tab Package\cr
Version: \tab 1.0.3\cr
Date: \tab 2020-02-21\cr
License: \tab GPL-2\cr
}

}
\author{Kelly Burkett}
\references{
Burkett KM, McNeney B, Graham J. Sampletrees and Rsampletrees: sampling gene genealogies conditional on SNP genotype data. Bioinformatics. 32:1580-2, 2016.
}

